Add C4-PlantUML Properties to Deployment_Nodes
Because of the way Deployment Nodes were rendered, the code path to add
C4-PlantUML AddProperty Attributes was not being called.

This change adds a check to if the `c4plantuml.elementProperties`
property has been set in the `startDeploymentNodeBoundary` method to
ensure any model properties assigned to Deployment Nodes will also be
rendered on exported C4-PlantUML diagrams.

This is an addition to the old behaviour where the properties were only rendered on container views.